Added interactive hotel management system with user, room, and booking services, including test scenario and menu-driven interface.

+++ src/main/java/com/entities/Booking.java
@@ -15,6 +15,6 @@ public class Booking {
     private  Room roomNumber;
     private  LocalDate checkIn;
     private  LocalDate checkOut;
-    private  int bookedPricePerNight;
+    private  Double bookedPricePerNight;
     private  int totalCost;
 }
